project/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
playwright
beautifulsoup4==4.12.2
pyngrok==7.0.0
requests==2.31.0
lxml==4.9.3
python-multipart==0.0.6
python-dotenv==1.0.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
pillow==10.1.0
pandas>=1.3.0
tqdm>=4.62.0
numpy>=1.20.0
ultralytics>=8.0.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-learn>=0.24.0
Pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
transformers>=4.11.0
timm>=0.4.12